==> NHANES/step3_impute_values.ipynb <==
# NHANES - Impute missing values

The purpose of this notebook is to impute missing values in variables of interest

**NOTE: Document code to explain what is happening at each level of the KNN, why 5 neighbors, what's happening with scale, fit, transform, etc.**

## Import packages

import pandas as pd
import numpy as np

# For missing value imputation
from sklearn.impute import KNNImputer
from sklearn.preprocessing import MinMaxScaler

## Load data

nhanes = pd.read_csv('data/nhanes.csv')

## Create model 

###  Model-Based Imputation with k-NN 

k-Nearest Neighbors (k-NN) method for imputation can effectively handle the nuances in the dataset by imputing missing 'ferritin' values based on the most similar respondents, considering a range of variables including 'race-ethnicity', 'household-income', and 'income-to-poverty-ratio'.

> - Preparation: Before imputation, it's essential to ensure that all predictor variables used in the k-NN algorithm are appropriately scaled. This ensures that each variable contributes equally to the distance calculation.
> - Choosing k: The choice of k (the number of neighbors) is crucial. A smaller k can make the imputation sensitive to noise, while a larger k might smooth out local variations too much. Cross-validation can help in selecting an optimal k value.
> - Imputation: Perform the k-NN imputation, using available variables as predictors to impute missing 'transformed_ferritin' values.

### Selecting columns for the imputation model

nhanes.loc[:, 'household-income'].fillna(nhanes['household-income'].median(), inplace=True)


nhanes.loc[:, 'income-to-poverty-ratio'].fillna(nhanes['income-to-poverty-ratio'].median(), 
                                                inplace=True)


nhanes.set_index('SEQN', inplace=True)

imputation_features = nhanes[['transformed_ferritin',
                              'tfr',
                              'race-ethnicity', 
                              'household-income', 
                              'income-to-poverty-ratio']]

imputation_features

### Scale features

# Features scaling to prepare for k-NN
scaler = MinMaxScaler()

imputation_features_scaled = pd.DataFrame(scaler.fit_transform(imputation_features), 
                                          columns=imputation_features.columns)

# Applying k-NN imputation
knn_imputer = KNNImputer(n_neighbors=5)

imputed_data = nhanes[['transformed_ferritin', 'tfr','race-ethnicity', 
                                   'household-income', 
                                   'income-to-poverty-ratio']].copy()


imputed_data

imputed_data = knn_imputer.fit_transform(imputation_features_scaled)

imputed_data_df = pd.DataFrame(imputed_data, columns=imputation_features.columns, 
                               index=imputation_features.index)

imputed_data_df

# Reverse scaling to get back to original scale for 'transformed_ferritin'
imputed_data_df = pd.DataFrame(scaler.inverse_transform(imputed_data_df), 
                               columns=imputation_features.columns,
                              index=imputation_features.index)

imputed_data_df

print(nhanes.index.is_unique)
print(imputed_data_df.index.is_unique)

imputed_data_df['transformed_ferritin'].isnull().sum()

imputed_data_df['tfr'].isnull().sum()

print(nhanes.index.equals(imputed_data_df.index))

nhanes = nhanes.join(imputed_data_df[['transformed_ferritin','tfr']], 
                     how='left', rsuffix='_imputed')

nhanes['transformed_ferritin_imputed'].isnull().sum()

nhanes['tfr_imputed'].isnull().sum()

# Displaying a summary to verify the imputation
nhanes[['transformed_ferritin', 'transformed_ferritin_imputed']].describe()

nhanes[['tfr', 'tfr_imputed']].describe()

nhanes['months-postpartum']=nhanes['months-postpartum'].fillna(0)

nhanes['months-postpartum'].isnull().sum()

## Save csv

nhanes.to_csv('data/nhanes.csv', index=False)
