==> NHANES/step4_construct_weights.ipynb <==
# NHANES - Construct 10 year sampling weights

Purpose of this notebook: **State the importance of sampling weights and why we will use MEC in our analyses**

## Import packages

import pandas as pd
import numpy as np

## Load data

nhanes = pd.read_csv('data/nhanes.csv')

## Construct sampling weights 
- For NHANES datasets, the use of sampling weights and sample design variables is necessary to obtain unbiased estimates and accurate standard errors and confidence intervals.
https://wwwn.cdc.gov/nchs/nhanes/tutorials/VarianceEstimation.aspx

- Constructing Weights when Combining Survey Cycles of Continuous NHANES, Combining 5 survey cycles (10 years)
https://wwwn.cdc.gov/nchs/nhanes/tutorials/weighting.aspx

#### Construct Weights Combining Survey Cycles of Continuous NHANES, Combining 5 survey cycles (10 years)


                                'WTINT2YR' = 'interview-weight',
                                'WTMEC2YR' = 'mec-exam-weight',

# SDDSRVYR represents the two-year data release cycle number
nhanes['SDDSRVYR'].value_counts()

nhanes['INT10YR']= 1/5 * nhanes['WTINT2YR']

nhanes['MEC10YR']= 1/5 * nhanes['WTMEC2YR']

## Save csv

nhanes.to_csv('data/nhanes.csv', index=False)
