==> NHANES/step5_descriptive_stats.ipynb <==
# NHANES - Descriptive statistics

Purpose of this notebook: Create "Table 1"

## Load libraries

library(dplyr)
library(ggplot2)
library(survey)
library(table1);

## Load data

nhanes <- read.csv("data/nhanes.csv")

## Set up survey design object

nhanes_design <- svydesign(ids = ~masked.variance.psu, 
                           strata = ~masked.variance.stratum, 
                           weights = ~MEC10YR,
                           nest=TRUE,
                           data = nhanes)

# View the survey design object
print(nhanes_design)

age_mean_estimate <- svymean(~age, design = nhanes_design)

mean_estimate
