==> NHANES/step6_prevalence_stats.ipynb <==
# NHANES - Prevalence statistics

## Load libraries

library(dplyr)
library(ggplot2)
library(survey)
library(table1);

## Load data

nhanes <- read.csv("data/nhanes.csv")

## Set up survey design object


design <- svydesign(ids = ~masked.variance.psu, 
                           strata = ~masked.variance.stratum, 
                           weights = ~MEC10YR,
                           nest = TRUE,
                           data = nhanes)

# View the survey design object
print(design)

## Calculate N & median values for TFR and FT by pregnancy_status

### TfR

#### N

# N for ID (TfR >= 8.3 mg/l) by pregnancy status
n_id_tfr_by_pregnancy <- nhanes %>%
  group_by(pregnancy.status) %>%
  summarise(ID_tfr = sum(ID_tfr == 1))

# Print the results
print(n_id_tfr_by_pregnancy)

#### Medians

# TfR medians by pregnancy status
medians <- svyby(~tfr, by = ~pregnancy.status, design, svyquantile, c(0.5), na.rm = TRUE)
print(medians)

# Medians by depression status and pregnancy status
medians_by_pregnancy_and_depression <- svyby(~tfr, by = ~pregnancy.status + depression, 
                                             design, svyquantile, c(0.5), na.rm = TRUE)

print(medians_by_pregnancy_and_depression)

#### Prevalence

# Calculate the prevalence of tfr >= 8.3 for each pregnancy status category
prevalence_by_pregnancy_status <- svyby(~ID_tfr, ~pregnancy.status, design, svymean, na.rm = TRUE)


# Print the results
print(prevalence_by_pregnancy_status)

prevalence_by_pregnancy_and_depression <- svyby(~ID_tfr, 
                                                ~pregnancy.status + depression, 
                                                design, svymean, na.rm = TRUE)


# Print the results
print(prevalence_by_pregnancy_and_depression)

### Rao Scott Chi-Square test

#### ID TfR >= 8.3 Prevalence

# Create a new design object for non-pregnant group (pregnancy.status == 0)
design_not_pregnant <- subset(design, pregnancy.status == 0)

# Not pregnant group: Perform the Rao-Scott chi-squared test
result <- svychisq(~ID_tfr + depression, design = design_not_pregnant)

# Print the test result
print(result)

# Create a new design object for pregnant group (pregnancy.status == 1)
design_pregnant <- subset(design, pregnancy.status == 1)

# Pregnant group: Perform the Rao-Scott chi-squared test
result <- svychisq(~ID_tfr + depression, design = design_pregnant)

# Print the test result
print(result)

# Create a new design object for postpartum group (pregnancy.status == 2)
design_postpartum <- subset(design, pregnancy.status == 2)

# Postpartum group: Perform the Rao-Scott chi-squared test
result <- svychisq(~ID_tfr + depression, design = design_postpartum)

# Print the test result
print(result)

### Ferritin

# N for low ferritin by pregnancy status 
n_id_ferr_by_pregnancy <- nhanes %>%
  group_by(pregnancy.status) %>%
  summarise(ID_ferr = sum(ID_ferr == 1))

# Print the results
print(n_id_ferr_by_pregnancy)

#### Medians

# ferritin medians by pregnancy status
medians_ferr <- svyby(~ferritin, by = ~pregnancy.status, design, svyquantile, c(0.5), na.rm = TRUE)
print(medians_ferr)

# Medians by depression status and pregnancy status
medians_by_pregnancy_and_depression_ferr <- svyby(~ferritin, by = ~pregnancy.status + depression, 
                                             design, svyquantile, c(0.5), na.rm = TRUE)

print(medians_by_pregnancy_and_depression_ferr)

#### Prevalence

# Calculate the prevalence of ID ferr  for each pregnancy status category
prevalence_by_pregnancy_status_ferr <- svyby(~ID_ferr, ~pregnancy.status, 
                                             design, svymean, na.rm = TRUE)
print(prevalence_by_pregnancy_status_ferr)

prevalence_by_pregnancy_and_depression_ferr <- svyby(~ID_ferr, 
                                                ~pregnancy.status + depression, 
                                                design, svymean, na.rm = TRUE)

print(prevalence_by_pregnancy_and_depression_ferr)

# Create a new design object for non-pregnant group (pregnancy.status == 0)
design_not_pregnant <- subset(design, pregnancy.status == 0)

# Not Pregnant group: Perform the Rao-Scott chi-squared test
result <- svychisq(~ID_ferr + depression, design = design_not_pregnant)

# Print the test result
print(result)

# Create a new design object for pregnant group (pregnancy.status == 1)
design_pregnant <- subset(design, pregnancy.status == 1)

# Pregnant group: Perform the Rao-Scott chi-squared test
result <- svychisq(~ID_ferr + depression, design = design_pregnant)

# Print the test result
print(result)

# Create a new design object for postpartum group (pregnancy.status == 2)
design_postpartum <- subset(design, pregnancy.status == 2)

# Postpartum group: Perform the Rao-Scott chi-squared test
result <- svychisq(~ID_ferr + depression, design = design_postpartum)

# Print the test result
print(result)
